update firebase init
